requirements.txt
accelerate==0.34.2
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anthropic==0.42.0
attrs==25.3.0
av==13.0.0
bchlib==2.1.3
bitsandbytes==0.45.0
black==24.1.0
blind-watermark==0.4.4
certifi==2025.8.3
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
Cython==3.1.3
DataProperty==1.0.1
datasets==2.16.1
decord==0.6.0
deepspeed==0.15.4
dill==0.3.7
docker-pycreds==0.4.0
einops==0.3.0
evaluate==0.4.3
fairscale==0.4.13
filelock==3.19.1
fire==0.7.0
flash_attn==2.8.3
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2023.10.0
ftfy==6.2.3
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
hf-xet==1.1.10
hf_transfer==0.1.8
hjson==3.1.0
httplib2==0.22.0
huggingface-hub==0.34.4
humanfriendly==10.0
identify==2.6.1
idna==3.10
imagenet-c==0.0.3
jieba==0.42.1
Jinja2==3.1.6
jiter==0.5.0
joblib==1.5.2
jsonlines==4.0.0
kiwisolver==1.4.9
kornia==0.6.0
libretranslatepy==2.1.1
lightning-utilities==0.11.7
# Editable install with no version control (llm-reasoners==1.0.2)
-e /hpc2hdd/JH_DATA/share/qzheng219/PrivateShareGroup/Expand/llm-reasoners
lmms_eval @ git+https://github.com/EvolvingLMMs-Lab/lmms-eval.git@0f1576ebbbcdb21c6d3bf46d0979396daa2e3ca9
loguru==0.7.2
lpips==0.1.4
lxml==6.0.1
markllm==0.1.5
MarkupSafe==3.0.2
matplotlib==3.10.6
mbstrdecoder==1.1.3
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.15
mutagen==1.47.0
networkx==3.5
ninja==1.11.1.1
nltk==3.9.1
nodeenv==1.9.1
numpy==2.3.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
openai==0.28.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
optimum==1.23.3
packaging==25.0
pandas==2.2.3
pathvalidate==3.2.1
peft==0.13.0
Pillow==9.4.0
portalocker==2.10.1
pre-commit==3.8.0
propcache==0.3.2
proto-plus==1.25.0
protobuf==5.29.2
psutil==7.0.0
pyarrow==21.0.0
pyarrow-hotfix==0.6
pycocoevalcap==1.2
pycocotools==2.0.8
pycryptodomex==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
pyDeprecate==0.3.1
pyparsing==3.2.4
pytablewriter==1.2.0
python-dateutil==2.9.0.post0
pytorch-lightning==1.4.2
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
rfc3986==1.5.0
sacrebleu==2.4.3
safetensors==0.4.5
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.0
sentencepiece==0.2.1
setproctitle==1.3.3
six==1.17.0
slack==0.0.2
slackclient==2.9.3
smmap==5.0.1
sqlitedict==2.1.0
sympy==1.14.0
tabledata==1.3.3
tabulate==0.9.0
tarski==0.8.2
tcolorpy==0.1.6
tenacity==8.3.0
threadpoolctl==3.6.0
tiktoken==0.7.0
timm==1.0.9
tokenizers==0.20.0
torch==2.8.0
torchmetrics==0.6.0
torchvision==0.23.0
tqdm==4.67.1
tqdm-multiprocess==0.0.11
transformers==4.45.1
transformers-stream-generator==0.0.5
translate==3.6.1
triton==3.4.0
typepy==1.3.2
typing_extensions==4.15.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.5.0
utils==1.0.2
Wand==0.6.13
wandb==0.18.2
wcwidth==0.2.13
websockets==13.1
xxhash==3.5.0
yarl==1.20.1
yt-dlp==2024.9.27
zss==1.2.0
